+++ tests/unit/test.base.js
@@ -1,0 +1,25 @@
+/*!
+ * @author:    Divio AG
+ * @copyright: http://www.divio.ch
+ */
+
+'use strict';
+/* global Cl, describe, it, expect, fixture, beforeEach, outdatedBrowser */
+
+// #############################################################################
+// UNIT TEST
+describe('base.js', function () {
+    beforeEach(function () {
+        fixture.setBase('tests/fixtures');
+        this.markup = fixture.load('outdatedBrowser.html');
+    });
+
+    it('loads Cl.Utils._document()', function () {
+        expect(window.Cl).toBeDefined();
+        expect(Cl.Utils).toBeDefined();
+    });
+
+    it('loads outdatedBrowser', function () {
+        expect(outdatedBrowser).toBeDefined();
+    });
+});